requirements.txt
django==1.9.4
daphne==0.9.3
channels==0.9.4
fake-factory==0.5.6
djangorestframework==3.3.2
httpie==0.9.3
